Neutron Trunk Port related fixes
- Add trunk details info to the port:

If the port is used as a trunk parent port
we need to add this information to the 'trunk_details' field.

- Remove vmi refs before deleting trunk port:

Delete references to subports and parent port.

- Fix trunk subport issues

- Add check for trunk sub_ports type:

Check for sub_ports type and throw exception if format is not correct.

- Set trunk admin_state_up according to the request:

If admin_state_up is passed in the create request we should
set it according to the request.

- Return "ACTIVE" status of trunk to openstack:

Do not use _port_ge_interface_status function to get status of trunk ports.
As there is no mechanism in TungstenFabric to properly operate with statuses
of ports, let's just always return ACTIVE.

Change-Id: I516946fce781932385d7c6abd4d71762bb8c803b