Add zookeeper lock on schema change
When TF services are trying to create keyspaces or tables
sometimes they make it concurrent. This can lead to the schema
inconsistency. To prevent this behaviour ZooKeeper lock
was implepented, it gives us confidence that at any snapshot
of the time only one client will do schema changes.
Also to prevent extra requests to Cassandra were added
functions to check whether this table/keyspace already created or not.

Depends-On: I6fa01cede7ece527129bfcdfb28ced6d1e8e85e1
Change-Id: I27f7c9cde8d7e5e077f7861fd1da176c650eaa79